Feature/install-singular-plugin (#8)
* chore: update ConfigSchema to use PluginSchema for plugins validation

* feat: Update plugin installation command to support installing a specific plugin

This commit modifies the plugin installation command to support installing a specific plugin by providing the plugin ID as an argument. The command now checks if a plugin ID is provided and installs only that plugin in the selected vaults. If no plugin ID is provided, it installs all plugins specified in the config file.

* chore: remove description constants